moved .h files to include/, fix src includes, moved tests to tests/
