+++ CMakeLists.txt
@@ -3,23 +3,20 @@ project(bedrock-viz VERSION 0.1.2)
 
 set(CMAKE_CXX_STANDARD 17)
 
-option(BEDROCK_VIZ_BUILD_TESTS "Build unit tests" OFF)
+option(BUILD_TESTS "Build unit tests" OFF)
 
 if (CMAKE_BUILD_TYPE STREQUAL Debug)
   add_compile_definitions(BEDROCK_VIZ_DEBUG)
 endif(CMAKE_BUILD_TYPE STREQUAL Debug)
 
-configure_file(src/config.h.in config.h)
+configure_file(include/config.h.in config.h)
 
 file(GLOB_RECURSE SRC_FILES src/*.cc)
-file(GLOB_RECURSE HEADER_FILES src/*.h)
-
-list(FILTER SRC_FILES EXCLUDE REGEX ".*_test.cc$")
+file(GLOB_RECURSE HEADER_FILES include/*.h)
+file(GLOB_RECURSE TEST_FILES tests/*.cc)
 
 add_executable(bedrock-viz
   ${SRC_FILES}
-
-  ${HEADER_FILES}
 )
 
 if(MSVC)
@@ -55,6 +52,7 @@ include_directories(
 )
 
 target_include_directories(bedrock-viz PRIVATE ${CMAKE_CURRENT_BINARY_DIR})
+target_include_directories(bedrock-viz PUBLIC ${PROJECT_SOURCE_DIR}/include)
 
 if (CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
   # GCC requires -lstdc++fs to use <filesystem>
@@ -79,7 +77,7 @@ install(DIRECTORY data/ DESTINATION share/bedrock-viz/data)
 install(DIRECTORY static/ DESTINATION share/bedrock-viz/static)
 
 # From LevelDB's CMakeLists.txt
-if(BEDROCK_VIZ_BUILD_TESTS)
+if(BUILD_TESTS)
   enable_testing()
   set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
   set(BUILD_GMOCK OFF CACHE INTERNAL "Don't bulid gmock")
@@ -88,14 +86,12 @@ if(BEDROCK_VIZ_BUILD_TESTS)
   add_subdirectory("third_party/googletest")
 
   file(GLOB_RECURSE SRC_FILES_NO_MAIN src/*.cc)
-  list(FILTER SRC_FILES_NO_MAIN EXCLUDE REGEX ".*_test.cc$")
   list(FILTER SRC_FILES_NO_MAIN EXCLUDE REGEX ".*/main.cc$")
 
   function(viz_test test_file)
     get_filename_component(test_target_name "${test_file}" NAME_WE)
     add_executable("${test_target_name}" 
       ${SRC_FILES_NO_MAIN}
-      ${HEADER_FILES}
       "${test_file}"
     )
 
@@ -118,10 +114,10 @@ if(BEDROCK_VIZ_BUILD_TESTS)
     endif()
 
     target_include_directories("${test_target_name}" PRIVATE ${CMAKE_CURRENT_BINARY_DIR})
+    target_include_directories("${test_target_name}" PUBLIC ${PROJECT_SOURCE_DIR}/include)
 
     add_test(NAME "${test_target_name}" COMMAND "${test_target_name}")
   endfunction(viz_test)
 
-  viz_test("src/example_test.cc")
-
-endif(BEDROCK_VIZ_BUILD_TESTS)
+  viz_test(${TEST_FILES})
+endif(BUILD_TESTS)
