+++ include/asset.h
