+++ include/global.h
