+++ include/xml/load_entity.h
