+++ src/minecraft/v2/biome.cc
@@ -1,10 +1,9 @@
-#include "biome.h"
+#include "minecraft/v2/biome.h"
+#include "config.h"
+#include "logger.h"
+#include "utils/pointer_array.h"
 
 #include <map>
-
-#include "config.h"
-#include "../../logger.h"
-#include "../../utils/pointer_array.h"
 
 namespace
 {
