+++ src/world/common.cc
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "world/common.h"
 
 namespace mcpe_viz {
     leveldb::ReadOptions levelDbReadOptions;
