+++ src/world/misc.cc
@@ -1,10 +1,10 @@
-#include "misc.h"
+#include "world/misc.h"
 
-#include "../define.h"
-#include "../nbt.h"
-#include "../logger.h"
-#include "../utils/unknown_recorder.h"
-#include "../minecraft/v2/block.h"
+#include "define.h"
+#include "nbt.h"
+#include "logger.h"
+#include "utils/unknown_recorder.h"
+#include "minecraft/v2/block.h"
 
 
 namespace mcpe_viz {
