+++ src/world/point_conversion.cc
@@ -1,6 +1,5 @@
-#include "point_conversion.h"
-
-#include "world.h"
+#include "world/point_conversion.h"
+#include "world/world.h"
 
 namespace mcpe_viz {
     void worldPointToImagePoint(int32_t dimId, double wx, double wz, double& ix, double& iy, bool geoJsonFlag) {
