+++ src/xml/load_block.cc
@@ -1,7 +1,7 @@
-#include "load_block.h"
-#include "../minecraft/v2/block.h"
-#include "../logger.h"
-#include "../util.h"
+#include "xml/load_block.h"
+#include "minecraft/v2/block.h"
+#include "logger.h"
+#include "util.h"
 
 namespace mcpe_viz
 {
